+++ packages/terra-tabs/src/common-tabs/_HiddenTab.jsx
@@ -9,6 +9,10 @@ import {
 } from 'keycode-js';
 import IconClose from 'terra-icon/lib/icon/IconClose';
 import { injectIntl } from 'react-intl';
+import { Draggable } from 'react-beautiful-dnd';
+import IconKnurling from 'terra-icon/lib/icon/IconKnurling';
+import { v4 as uuidv4 } from 'uuid';
+import VisuallyHiddenText from 'terra-visually-hidden-text';
 import {
   enableFocusStyles,
   disableFocusStyles,
@@ -101,6 +105,15 @@ const propTypes = {
    * Indicates if the pane should be disabled.
    */
   isDisabled: PropTypes.bool,
+  /**
+   * Whether or not the tab is draggable.
+   */
+  isDraggable: PropTypes.bool,
+  /**
+   * @private
+   * intl object programmatically imported through injectIntl from react-intl.
+   */
+  intl: PropTypes.shape({ formatMessage: PropTypes.func }).isRequired,
 };
 
 const defaultProps = {
@@ -109,6 +122,7 @@ const defaultProps = {
   showAddButton: false,
   isDisabled: false,
   isClosable: false,
+  isDraggable: false,
 };
 
 const HiddenTab = ({
@@ -131,6 +145,8 @@ const HiddenTab = ({
   isClosable,
   intl,
   isDisabled,
+  isDraggable,
+  intl,
 }) => {
   const attributes = {};
   const theme = React.useContext(ThemeContext);
@@ -166,7 +182,7 @@ const HiddenTab = ({
     });
   }
   const onKeyDown = (event) => {
-    if (event.nativeEvent.keyCode === KEY_RETURN || event.nativeEvent.keyCode === KEY_SPACE) {
+    if (event.nativeEvent.keyCode === KEY_RETURN || (event.nativeEvent.keyCode === KEY_SPACE && !isDraggable)) {
       handleOnSelect(event);
       if (onChange) {
         onChange(event, itemKey);
@@ -177,7 +193,8 @@ const HiddenTab = ({
       event.stopPropagation();
       onCloseClick(event);
     } else {
-      handleArrows(event, index, tabIds);
+      const isDragging = !!document.querySelectorAll('[data-terra-menu-drag-focus="true"]').length && isDraggable;
+      handleArrows(event, index, tabIds, isDragging);
     }
   };
 
@@ -197,6 +214,36 @@ const HiddenTab = ({
   attributes.onMouseDown = disableFocusStyles;
   attributes['data-focus-styles-enabled'] = true;
   attributes['aria-selected'] = isSelected;
+
+  const onFocusResponse = intl.formatMessage({ id: 'Terra.tabs.focus' });
+  const responseId = `terra-hidden-tab-pane-response=${uuidv4()}`;
+
+  if (isDraggable) {
+    return (
+      <Draggable key={id} draggableId={id} index={index}>
+        {(provided) => (
+          <div
+            {...attributes}
+            ref={provided.innerRef}
+            {...provided.draggableProps}
+            {...provided.dragHandleProps}
+            id={id}
+            aria-controls={associatedPanelId}
+            role="tab"
+            className={hiddenClassNames}
+            aria-disabled={isDisabled}
+            aria-describedby={responseId}
+          >
+            <VisuallyHiddenText aria-hidden id={responseId} text={onFocusResponse} />
+            <IconKnurling />
+            <div className={cx('checkbox')}>{isSelected ? <IconCheckmark /> : null}</div>
+            {showIcon && <div>{icon}</div>}
+            <div className={cx('label', { 'with-icon': showIcon })}>{label}</div>
+          </div>
+        )}
+      </Draggable>
+    );
+  }
 
   return (
     <div