+++ packages/terra-tabs/src/common-tabs/_Tab.jsx
@@ -222,8 +222,7 @@ const Tab = ({
   attributes['aria-selected'] = isSelected;
   attributes.style = { zIndex };
 
-  // TBD: Add Translations
-  const onFocusResponse = 'Press Enter to activate a tab, or Press space bar to start a drag. When dragging you can use the arrow keys to move the item around, space bar to drop and escape to cancel. Ensure your screen reader is in focus mode or forms mode';
+  const onFocusResponse = intl.formatMessage({ id: 'Terra.tabs.focus' });
   const responseId = `terra-tab-pane-response=${uuidv4()}`;
 
   if (isDraggable) {
@@ -245,12 +244,11 @@ const Tab = ({
             data-terra-tabs-show-focus-styles
             data-terra-tab-draggable
           >
-            <div>
-
-              <IconKnurling className={cx('icon-knurling')} />
+            <div className={variant === 'framework' ? cy('inner', 'draggable-inner') : cx('inner')}>
+              <IconKnurling />
               <VisuallyHiddenText aria-hidden id={responseId} text={onFocusResponse} />
               {customDisplay || icon}
-              {(!customDisplay && !isIconOnly) && <span className={cx('label')}>{label}</span>}
+              {(!customDisplay && !isIconOnly) && <span className={variant === 'framework' ? cy('label') : cx('label')}>{label}</span>}
             </div>
             {isClosable && (
             <div